Looking how to best implement Problem
Dumb typo with Problem's passed object.

Also single list is way easier than 2d list for this.

While the syntax for 2d might be compact,
it isn't flat. Since problem action and result is in one place, a larger
but flatter implementation might be easier on the eyes...and more sane.